Adding a click handler to the delete buttons to remove the associated todoItem.
Currently, this feature is only cosmetic and is only reflected in the UI.
The todoItem still exists within the todo.list array.
I would like to avoid removing items from the array entirely to allow a history function to be added later,
however, a state marker needs to be added to the todoItem object to keep the view in sync with the data model.